Python

+++ Entrada_datos/entrada_datos.py
@@ -1,31 +1,50 @@
 # Entrada de datos
 
+
 nombre = input("Introduce el nombre: ")
+
 print(f'El valor es: {nombre}')
+
 
 # Pedir edad
 
+
 edad = int(input("Introduce la edad: "))
+
 print(f'edad es: {edad}')
+
 
 # Ejercicio sistema de empleados
 
+
 Nombre = input("Introduce el nombre del empleado: ")
+
 Edad = int(input("Introduce la edad: "))
-Salario = float(input("Introduce el salario: "))
+
+Salario = int(input("Introduce el salario: "))
+
 Es_jefe = input("Es jefe de departamento?")
+
 
 # Convertir a bool la variable jefe
 
+
 Es_jefe = Es_jefe.lower() == 'si'
+
 
 #Imprimir valores
 
+
 print(f' Datos de empleado: ')
+
 print(f'Nombre: {Nombre}')
+
 print(f'Edad: {Edad}')
+
 print(f'Salario: {Salario}')
+
 print(f'Jefe de departamento: {Es_jefe}')
+
